requirements.txt
streamlit_lottie==0.0.2
streamlit==1.3.0
requests==2.24.0
Pillow==8.4.0
streamlit==0.71.0
pandas==1.1.3
scikit-learn==0.23.2
numpy==1.19.2
